Track is responsible for playing itself
Since players and speakers get cleared and reinitialized each time
a track is stopped or started, might as well do it in the Track itself
and keep the logic as local as possible.

In moving the logic over, also move towards a Promise based model for
async making-things-happen-ness.

There's a lot of extraneous code. This is just the first commit where
everything works without exploding.