Undo feature
Enqueuing track or tracks adds an undo level (max 5) which can be
rewound with the 'u' key. Only enqueuing adds undos at the moment.

Function in a closure changed to a View prototype method so it could
be passed into the UndoBuffer as the callback to perform upon rewinding
the undo layers.